Terminal
I thought it might be a cool idea to make it a little bit more user
friendly by making it open up a Terminal when the .jar file is double
clicked. It is not fully finished, and the user input would have to
change to comply with it(I think). See if you like, and if not that’s ok